mantle/ore: make ARM versions of Azure commands primary
According to @bgilbert ARM is the new API and ASM is the old one.
The mantle ASM code is left over from Container Linux and no one
should be using it at this point because it uses some unpublished
Azure APIs. Let's make the "arm" versions of our commands the
primary versions and delete the old code. This means:

- "create-image-arm" now becomes "create-image"
- "delete-image-arm" now becomes "delete-image"
- "upload-blob-arm" now becomes "upload-blog"

The "-arm" versions of the commands are aliased to allow consumers to
keep calling it that way for now.

(cherry picked from commit 0226b078850fdad7e1e08321a8fd5a25b619b9c8)
